make code example in introduction doc self-contained
This code snippet is referenced in the introduction documentation of akka-http:
https://doc.akka.io/docs/akka-http/current/introduction.html
It was a frustrating experience (again) with examples from the doc. I had to spend some time to find out why the code snippet didn't compile. Mentioning this required dependency for this snippet would save me and others time.